Revert "NRNF-68 first try to extract achievements to own component"
This reverts commit 37286d32
